complete photo upload

+++ client/src/Components/MyArticles.jsx
@@ -1,8 +1,12 @@
+//========================== Import Modules Start ===========================
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { NavLink } from 'react-router-dom';
 import { deleteArticle, userProfile } from '../Actions/actions';
+//========================== Import Modules End =============================
 
+
+//============================= My Articles Component Start =============================
 const MyArticles = () => {
 
     const User = useSelector(state => state.User);
@@ -11,7 +15,9 @@ const MyArticles = () => {
     const dispatch = useDispatch();
 
     const handleDelete = (id) => {
-        dispatch(deleteArticle(id))
+        if(window.confirm('Are You Sure!')){
+            dispatch(deleteArticle(id))
+        }
     }
     
     useEffect(() => {
@@ -28,7 +34,7 @@ const MyArticles = () => {
                 User &&  User.Articles && User.Articles.map((article) => {                 
                     return(
                         <>
-                            <div class="profile">
+                            <div class="profile" key={article._id}>
                                 <div className='subDiv'>                               
                                     <label>Title </label>
                                     <p>{article.title}</p>
@@ -39,7 +45,7 @@ const MyArticles = () => {
                                     <label>Description </label>
                                     <p>{article.description}</p>
                                 </div>
-                                <div className='img'>
+                                <div className='banner'>
                                     <img src={article.banner} alt='article banner'/>
                                 </div>
                                 <NavLink to = {`/EditArticle/:?id=${article._id}`}><button>Edit</button></NavLink>
@@ -54,4 +60,8 @@ const MyArticles = () => {
   )
 }
 
+//============================= Export Default Start =============================
+
 export default MyArticles;
+
+//============================= Export Default End =============================
